requirements/development.txt
-r base.txt

# Development specific packages
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
coverage==7.3.2
pre-commit==3.5.0